+++ app/src/main/java/org/oxycblt/auxio/util/PrimitiveUtil.kt
@@ -20,6 +20,7 @@ package org.oxycblt.auxio.util
 import android.database.Cursor
 import android.database.sqlite.SQLiteDatabase
 import android.os.Looper
+import androidx.core.math.MathUtils
 import androidx.fragment.app.Fragment
 import org.oxycblt.auxio.BuildConfig
 
@@ -51,3 +52,9 @@ fun <T> unlikelyToBeNull(value: T?): T {
 
 /** Require the fragment is attached to an activity. */
 fun Fragment.requireAttached() = check(!isDetached) { "Fragment is detached from activity" }
+
+fun Int.clamp(min: Int, max: Int): Int = MathUtils.clamp(this, min, max)
+
+fun Long.clamp(min: Long, max: Long): Long = MathUtils.clamp(this, min, max)
+
+fun Float.clamp(min: Float, max: Float): Float = MathUtils.clamp(this, min, max)